# back/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Una API para gestionar los usuarios, pedidos y productos de Delilah Restó",
    "version": "4.0.0",
    "title": "Delilah Restó - Sprint #4",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "nicolax0416@gmail.com"
    }
  },"tags": [
...
  "schemes": ["http","https"],
  "paths": {
...
    "/api/v1/paypal/payment/{orderId}": {
      "post": {
        "tags": ["plataformas de pago"],
        "summary": "Crear la orden de pago por Paypal",
        "description": "Endpoint que sirve para poder crear la ordenes de pago a través de la plataforma Paypal",
        "produces": ["application/json"],
        "parameters": [{
          "name": "authentication",
          "in": "header",
          "description": "authentication generado en /api/v1/login",
          "required": true,
          "type": "string"
        },{
          "name": "orderId",
          "in": "path",
          "description": "El id de la orden",
          "required": true,
          "type": "number"
        },{
          "name": "body",
          "in": "body",
          "required": true,
          "description": "Se debe mandar el monto total de la orden",
          "schema": {
            "example": {
              "amount": 1200
            }
          }
        }],
        "responses": {
          "200": {
            "description": "Operación correcta"
          },
          "401": {
            "description": "Acceso Denegado"
          },
          "403": {
            "description": "Operación Prohibida"
          },
          "406":{
            "description": "Se requiere el Header authentication"
          },
          "500": {
            "description": "Error"
          }
        }
      }
    },
    "/api/v1/paypal/redirect/{orderId}": {
      "get": {
        "tags": ["plataformas de pago"],
        "summary": "Pago exitoso por Paypal",
        "produces": ["application/json"],
        "description": "Endpoint que sirve para poder cambiar el status de la orderId a 'Payment completed. Preparing Order' a través de la redirección exitosa del pago de Paypal. NOTA: Es de uso exclusivo del Front",
        "parameters": [{
          "name": "token",
          "in": "query",
          "description": "Este parametro es proporcionado por la redirección del pago exitoso de la plataforma Paypal",
          "required": true,
          "type": "string"
        },
        {
          "name": "PayerId",
          "in": "query",
          "description": "Este parametro es proporcionado por la redirección del pago exitoso de la plataforma Paypal",
          "required": true,
          "type": "string"
        },{
          "name": "orderId",
          "in": "path",
          "description": "El id de la orden",
          "required": true,
          "type": "number"
        }],
        "responses": {
          "200": {
            "description": "The orderId XXXX was successfully paid."
        },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Invalid Token or PayerId"
          }
        }
      }
    },
    "/api/v1/paypal/cancel/{orderId}": {
      "get": {
        "tags": ["plataformas de pago"],
        "summary": "Pago exitoso por Paypal",
        "produces": ["application/json"],
        "parameters": [{
            "name": "orderId",
            "in": "path",
            "description": "El id de la orden",
            "required": true,
            "type": "number"
          }
        ],
        "description": "Endpoint que sirve para poder cambiar el status de la orderId a 'Payment unsuccessful' a través de la redirección exitosa del pago de Paypal. NOTA: Es de uso exclusivo del Front",
        "responses": {
          "200": {
            "description": "The orderId was unsuccessfully paid."
          },
          "404": {
            "description": "The orderId was not found."
          }
        }
      }
    },
    "/api/v1//mercadopago/payment/{orderId}": {
      "post": {
        "tags": ["plataformas de pago"],
        "summary": "Crear la orden de pago por Mercado Pago",
        "produces": ["application/json"],
        "parameters": [{
          "name": "authentication",
          "in": "header",
          "description": "authentication generado en /api/v1/login",
          "required": true,
          "type": "string"
        },{
          "name": "orderId",
          "in": "path",
          "description": "El id de la orden",
          "required": true,
          "type": "number"
        }],
        "responses": {
          "200": {
            "description": "Operación correcta"
          },
          "401": {
            "description": "Acceso Denegado"
          },
          "403": {
            "description": "Operación Prohibida"
          },
          "406":{
            "description": "Se requiere el Header authentication"
          },
          "500": {
            "description": "Error"
          }
        }
      }
    },
    "/api/v1/mercadopago/success/{orderId}": {
      "get": {
        "tags": ["plataformas de pago"],
        "summary": "Pago exitoso por Mercado Pago",
        "produces": ["application/json"],
        "description": "Endpoint que sirve para poder cambiar el status de la orderId a 'Payment completed. Preparing Order' a través de la redirección exitosa del pago de Mercado Pago. NOTA: Es de uso exclusivo del Front",
        "parameters": [{
          "name": "orderId",
          "in": "path",
          "description": "El id de la orden",
          "required": true,
          "type": "number"
        }],
        "responses": {
          "200": {
            "description": "The orderId was successfully paid."
        },
          "404": {
            "description": "Order not found"
          },
          "500": {
            "description": "Error"
          }
        }
      }
    },
    "/api/v1/mercadopago/pending/{orderId}": {
      "get": {
        "tags": ["plataformas de pago"],
        "summary": "Pago pendiente por Mercado Pago",
        "produces": ["application/json"],
        "parameters": [{
            "name": "orderId",
            "in": "path",
            "description": "El id de la orden",
            "required": true,
            "type": "number"
          }
        ],
        "description": "Endpoint que sirve para poder cambiar el status de la orderId a 'Pending for the payment' a través de la redirección del pago pendiente de Mercado Pago. NOTA: Es de uso exclusivo del Front",
        "responses": {
          "200": {
            "description": "The orderId was unsuccessfully paid."
          },
          "404": {
            "description": "The orderId was not found."
          },
          "500": {
            "description": "Error"
          }
        }
      }
    },
    "/api/v1/mercadopago/failure/{orderId}": {
      "get": {
        "tags": ["plataformas de pago"],
        "summary": "Pago rechazado por Mercado Pago",
        "produces": ["application/json"],
        "parameters": [{
            "name": "orderId",
            "in": "path",
            "description": "El id de la orden",
            "required": true,
            "type": "number"
          }
        ],
        "description": "Endpoint que sirve para poder cambiar el status de la orderId a 'Payment unsuccessful' a través de la redirección del pago fallido de Mercado Pago. NOTA: Es de uso exclusivo del Front",
        "responses": {
          "200": {
            "description": "The orderId was unsuccessfully paid."
          },
          "404": {
            "description": "The orderId was not found."
          },
          "500": {
            "description": "Error"
          }
        }
      }
    }
  }
}